requirements.txt
flask>=3.0
opencv-python>=4.9
anthropic>=0.40
python-dotenv>=1.0
